+++ src/index.ts
@@ -1,11 +1,7 @@
 import Logger from '@progressive-victory/logger';
 
-export { ExtendedClient as Client } from './Classes/ExtendedClient';
-
-export type { ExtendedClientOptions as ClientOptions } from './util';
-
 export {
-	ChatInputCommand,
+	ChatInputCommand, ExtendedClient as Client,
 	Command,
 	ContextMenuCommand,
 	Event,
@@ -18,8 +14,6 @@ export {
 
 export { ExtraColor, TimeStyles, tsNodeRun } from './util';
 
-export type { LocalizedHelpInfo } from './util';
-
-export type { TimeStyle } from './util';
+export type { ExtendedClientOptions as ClientOptions, LocalizedHelpInfo, TimeStyle } from './util';
 
 export const logger = new Logger();
